// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5";
  };
  public: {
    Tables: {
// ...
      skill_categories: {
        Row: {
          created_at: string;
          icon: string | null;
          id: string;
          name: string;
          order_index: number | null;
        };
        Insert: {
          created_at?: string;
          icon?: string | null;
          id?: string;
          name: string;
          order_index?: number | null;
        };
        Update: {
          created_at?: string;
          icon?: string | null;
          id?: string;
          name?: string;
          order_index?: number | null;
        };
        Relationships: [];
      };
      skills: {
        Row: {
          category_id: string | null;
          created_at: string;
          icon: string | null;
          id: string;
          level: number | null;
          name: string;
          order_index: number | null;
          svg_url: string | null;
          svg_url_dark: string | null;
        };
        Insert: {
          category_id?: string | null;
          created_at?: string;
          icon?: string | null;
          id?: string;
          level?: number | null;
          name: string;
          order_index?: number | null;
          svg_url?: string | null;
          svg_url_dark?: string | null;
        };
        Update: {
          category_id?: string | null;
          created_at?: string;
          icon?: string | null;
          id?: string;
          level?: number | null;
          name?: string;
          order_index?: number | null;
          svg_url?: string | null;
          svg_url_dark?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "skills_category_id_fkey";
            columns: ["category_id"];
            isOneToOne: false;
            referencedRelation: "skill_categories";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
};